button-wrap: Support Block API Version 2
