+++ blocks/src/block/button-wrap/index.php
@@ -1,0 +1,11 @@
+<?php
+
+if( function_exists('register_block_type_from_metadata')) {
+  add_action( 'init', 'register_block_cocoon_button_wrap', 99 );
+  function register_block_cocoon_button_wrap() {
+    register_block_type_from_metadata(
+      __DIR__,
+      array()
+    );
+  }
+}
