+++ blocks/src/blocks.js
@@ -16,6 +16,7 @@ import * as balloon from './block/balloon';
 import * as blankBox from './block/blank-box';
 import * as blogCard from './block/blogcard';
 import * as button from './block/button';
+import * as buttonWrap from './block/button-wrap';
 import * as iconBox from './block/icon-box';
 import * as iconList from './block/icon-list';
 import * as infoBox from './block/info-box';
@@ -27,6 +28,7 @@ const cocoonBlocks = [
   blankBox,
   blogCard,
   button,
+  buttonWrap,
   iconBox,
   iconList,
   infoBox,
@@ -83,7 +85,7 @@ registerCocoonBlocks();
 //import './block/balloon-ex/block.js';
 //import './block/blogcard/block.js';
 //import './block/button/block.js';
-import './block/button-wrap/block.js';
+//import './block/button-wrap/block.js';
 import './block/toggle-box/block.js';
 import './block/search-box/block.js';
 import './block/timeline/block.js';
