+++ blocks/src/init.php
@@ -279,6 +279,7 @@ require_once abspath(__FILE__).'block/balloon/index.php';
 require_once abspath(__FILE__).'block/blank-box/index.php';
 require_once abspath(__FILE__).'block/blogcard/index.php';
 require_once abspath(__FILE__).'block/button/index.php';
+require_once abspath(__FILE__).'block/button-wrap/index.php';
 require_once abspath(__FILE__).'block/icon-box/index.php';
 require_once abspath(__FILE__).'block/icon-list/index.php';
 require_once abspath(__FILE__).'block/info-box/index.php';
